grab_info deleted. modis metadata added
the grab_info object was left over from a very early attempt to
establish filename based metadata retrieval. It's only remaining use was
in the modis mosaic function, but a specific modis_metadata class has
been added to directly replace the existing grab_info function. Its much
cleaner. Also moved landsat_metadata into its own module so dnppy is now
internally consistent with all metadata handling.
